requirements.txt
fastapi
uvicorn[standard]
sqlmodel
pydantic
python-dotenv
google-generativeai
instructor
openai